backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
pycryptodome==3.20.0
numpy==2.0.1
pydantic==2.8.2
